requirements.txt
Django==2.0.3
djangorestframework==3.7.7
mysqlclient==2.0.1
stripe==2.51.0
